Revamped the UI
This will be the last update for this I need to get back to my game. If you have any comments or concerns you can message me on discord at Defiant Program#2410. If there's any bugs that I can fix I'll do a 1.31 or something in case I missed something. 
New boxlist that will let you copy a single file to multiple folders, only one at a time.
New dropdown that will auto populate the folder location to the default C:\ location of HK, select Desktop to turn this off in the dropdown
New help button to clean up the UI
Lots of spaghetti to make it work with the new features.
This is an open source and open license program, I made this to help the speeder community. Thank you for this opportunity to experiment I had never programmed without additional built in libraries like Unity. If anyone wants to fork this or expand it to support more games you're more than welcome to edit the files or use this as a template to make a new better one.